Move related media to the sidebar
