+++ app/components/media/media-information.js
@@ -1,6 +1,6 @@
 import Component from 'ember-component';
 import get from 'ember-metal/get';
-import computed, { not } from 'ember-computed';
+import computed, { equal, not } from 'ember-computed';
 import humanizeDuration from 'client/utils/humanize-duration';
 import moment from 'moment';
 
@@ -18,10 +18,8 @@ export default Component.extend({
   licensors: computedProduction('licensor'),
   studios: computedProduction('studio'),
 
-  isAnime: computed('media', function() {
-    return get(this, 'media.modelType') === 'anime';
-  }).readOnly(),
-  isManga: not('isAnime').readOnly(),
+  isAnime: equal('media.modelType', 'anime'),
+  isManga: not('isAnime'),
 
   season: computed('media.startDate', function() {
     const start = get(this, 'media.startDate');
