+++ app/styles/layout/_sidebars.scss
@@ -187,7 +187,7 @@
   }
 }
 
-.favorite-characters-panel {
+.favorite-characters-panel, .media--related {
   .favorite-item a {
     width: 100%;
   }
