+++ app/styles/pages/_media-pages.scss
@@ -208,6 +208,7 @@
 }
 
 .media--reactions {
+  margin-bottom: 20px;
   .reactions-header {
     margin-bottom: 15px;
     h5 {
@@ -360,6 +361,10 @@
   background: $foreground-background-color;
   border: 1px solid darken($background-color, 5);
   border-radius: 3px;
+  .more-link {
+    font-size: 12px;
+    color: $body-secondary-text-color;
+  }
 }
 
 .media-sidebar {
@@ -601,5 +606,12 @@
     height: 38px;
     float: left;
     margin-right: 5px;
+  }
+}
+
+.media--feed {
+  margin-top: 20px;
+  h5 {
+    margin-bottom: 20px;
   }
 }
