--- app/templates/components/media/media-relationships.hbs
@@ -1,23 +1,0 @@
-<h5>{{t "media.show.summary.relationships.header"}}</h5>
-{{#each-in relationships as |type medias|}}
-  <div class="related--panel">
-    <div class="related--wrapper">
-      <div class="related--title">{{t (concat "media.show.summary.relationships." type)}}:</div>
-      <div class="related--posters">
-        {{#each medias as |media|}}
-          <div class="related--item">
-            {{#media-popover
-              attachment="bottom middle"
-              targetAttachment="top middle"
-              media=media
-            }}
-              <a href={{href-to (concat media.modelType ".show") media.slug}}>
-                {{lazy-image src=(image media.posterImage "small")}}
-              </a>
-            {{/media-popover}}
-          </div>
-        {{/each}}
-      </div>
-    </div>
-  </div>
-{{/each-in}}
