+++ app/templates/components/media/media-summary.hbs
@@ -2,8 +2,14 @@
   {{! Media Information }}
   {{media/media-information media=media}}
 
+  {{! Media Attributes }}
+  {{! @TODO }}
+
   {{! Featured Characters }}
   {{media/featured-characters media=media}}
+
+  {{! Related Media }}
+  {{media/media-related media=media}}
 </div>
 
 {{! Kitsu Media Summary Sidebar }}
