+++ translations/en-us.yaml
@@ -188,7 +188,8 @@ media:
             other {}}"
       categories: "Categories"
       relationships:
-        header: "Franchise"
+        header: "More from this Series"
+        more: "View all in {title}"
         sequel: "Sequel"
         prequel: "Prequel"
         alternative_setting: "Alt. setting"
